refactor(sessions): split session-domain types into sessions/types
Phase 2 of 4 toward issue #181. Move SessionStatus, Session,
AgentActivity (and FileChange/ToolCall/TestResult/TestFailure/
ContextWindowStatus/UsageMetrics) to src/features/sessions/types/index.ts.
workspace/types/index.ts retains NavigationItem, Project,
ContextPanelType, ContextPanelState, Terminal, WorkspaceState — Project
imports Session from sessions/types for its sessions field.

Update ~20 importers across workspace/ and sessions/ to point at the
new location; sessions/ subtree now uses '../types' (own dir) instead
of the Phase-1 transitional '../../workspace/types'.

Refs #181
