+++ src/features/sessions/components/Tab.tsx
@@ -1,5 +1,5 @@
 import { type ReactElement, type KeyboardEvent } from 'react'
-import type { Session } from '../../workspace/types'
+import type { Session } from '../types'
 import type { Agent } from '../../../agents/registry'
 import { StatusDot } from '../../workspace/components/StatusDot'
