+++ src/features/sessions/utils/lineDelta.test.ts
@@ -1,6 +1,6 @@
 import { describe, test, expect } from 'vitest'
 import { lineDelta } from './lineDelta'
-import type { Session } from '../../workspace/types'
+import type { Session } from '../types'
 
 const sessionWith = (
   fileChanges: Session['activity']['fileChanges']
