+++ src/features/sessions/utils/lineDelta.ts
@@ -1,4 +1,4 @@
-import type { Session } from '../../workspace/types'
+import type { Session } from '../types'
 
 export const lineDelta = (
   session: Session
