+++ src/features/sessions/utils/statePill.ts
@@ -1,4 +1,4 @@
-import type { Session } from '../../workspace/types'
+import type { Session } from '../types'
 
 export const STATE_PILL_LABEL: Record<Session['status'], string> = {
   running: 'running',
