+++ src/features/workspace/utils/pickNextVisibleSessionId.test.ts
@@ -3,7 +3,7 @@ import {
   isOpenSessionStatus,
   pickNextVisibleSessionId,
 } from './pickNextVisibleSessionId'
-import type { Session, SessionStatus } from '../types'
+import type { Session, SessionStatus } from '../../sessions/types'
 
 const buildSession = (id: string, status: SessionStatus): Session => ({
   id,
